Made WebServer more reusable
Only non-general (non-reusable) part of WebServer was the device reboot after the submitted configuration had been stored.
Removed this from the WebServer and added it (via a function reference call) to Basecamp.

Added a reset function, which clears all interface elements such that new elemets can be added at will.
The reset function can be used if the web server should be reconfigured during runtime. But it might not work yet.